Added the new version of PDF parser (v2)
This new version of pdf parser implements functionality for merging data frames based on column names. There are some bugs: first off the column names are always numbered, maybe we can simply remove the column names and replace them. Another bug is when the data frames are merged, the real column names are not, so it it simply two datasets on top of one another. Hopefully, you (kevin) can fix this